+++ packages/xchain-mayamidgard-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [9229e99]
+  - @xchainjs/xchain-mayamidgard@0.1.1
+
 ## 0.1.4
 
 ### Patch Changes