+++ packages/xchain-mayamidgard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.1
+
+### Patch Changes
+
+- 9229e99: Update apis to the latest specs
+
 ## v0.1.0 (2023-11-24)
 
 ### Update