+++ packages/xchain-mayanode/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.3
+
+### Patch Changes
+
+- 9229e99: Update apis to the latest specs
+
 ## v0.1.2 (2023-11-14)
 
 ### Update