+++ packages/xchain-thorchain-amm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.8.15
+
+### Patch Changes
+
+- Updated dependencies [9229e99]
+  - @xchainjs/xchain-thorchain-query@0.7.3
+  - @xchainjs/xchain-thornode@0.3.10
+
+## 0.8.14
+
+### Patch Changes
+
+- Updated dependencies [04c87ae]
+  - @xchainjs/xchain-thorchain-query@0.7.2
+
 ## 0.8.13
 
 ### Patch Changes