+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.7.3
+
+### Patch Changes
+
+- 9229e99: Update apis to the latest specs
+- Updated dependencies [9229e99]
+  - @xchainjs/xchain-thornode@0.3.10
+
+## 0.7.2
+
+### Patch Changes
+
+- 04c87ae: Native decimal bug fix
+
 ## 0.7.1
 
 ### Patch Changes